Refactor input handling with strategy pattern
